fix: change joinDto directory to AuthDto
